Kernel+LibC: Use 64 bit values for ino_t
Since the InodeIndex encapsulates a 64 bit value, it is correct to
ensure that the Kernel is exposing the entire value and the LibC is
aware of it.

This commit requires an entire re-compile because it's essentially a
change in the Kernel ABI, together with a corresponding change in LibC.

+++ Kernel/UnixTypes.h
@@ -357,7 +357,7 @@ enum {
 #define TCSAFLUSH 2
 
 typedef u32 dev_t;
-typedef u32 ino_t;
+typedef u64 ino_t;
 typedef u16 mode_t;
 typedef u32 nlink_t;
 typedef u32 uid_t;
