+++ Userland/Libraries/LibC/sys/types.h
@@ -32,7 +32,7 @@ typedef char* caddr_t;
 
 typedef int id_t;
 
-typedef uint32_t ino_t;
+typedef uint64_t ino_t;
 typedef int64_t off_t;
 
 typedef uint32_t blkcnt_t;
